back/requirements.txt
gunicorn==19.9.0
chardet==2.3.0
Django==2.2.4
django-webpack-loader==0.6.0
docker==3.3.0
docker-compose==1.21.2
docx2txt==0.6
geotext==0.4.0
py==1.7.0
pytz==2018.4
regex==2019.2.3
Unidecode==1.0.23
uwsgi>=2.0,<2.1
django-cors-headers==3.0.2
django-import-export==1.2.0
